add all http api

+++ cmd/gophermart/migrations/00002_create_orders_table.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS orders (
     order_id bigint not null PRIMARY KEY,
     user_id bigint not null,
     status order_status_type,
-    accrual integer,
+    accrual double precision,
     uploaded_at timestamptz not null DEFAULT NOW(),
     changed_at timestamptz not null DEFAULT NOW()
 );
